[TMail] refactor: group all common configurations and tools in one folder
