+++ tmail-backend/tmail-third-party/ai-bot/src/main/java/com/linagora/tmail/aibotreplier/AIBotMailet.java
@@ -51,7 +51,7 @@ import com.github.fge.lambdas.Throwing;
 import com.google.common.base.Preconditions;
 import com.google.common.base.Strings;
 import com.google.common.collect.ImmutableList;
-
+import com.linagora.tmail.common.chatlanguagemodel.LlmConfig;
 
 import dev.langchain4j.data.message.ChatMessage;
 import dev.langchain4j.data.message.SystemMessage;
