+++ tmail-backend/tmail-third-party/ai-bot/src/main/java/com/linagora/tmail/common/chatlanguagemodel/LlmConfig.java
@@ -16,7 +16,7 @@
  *  more details.                                                   *
  ********************************************************************/
 
-package com.linagora.tmail.mailet;
+package com.linagora.tmail.common.chatlanguagemodel;
 
 import java.net.MalformedURLException;
 import java.net.URI;
