+++ tmail-backend/tmail-third-party/ai-bot/src/main/java/com/linagora/tmail/common/chatlanguagemodel/LlmModel.java
@@ -16,33 +16,7 @@
  *  more details.                                                   *
  ********************************************************************/
 
-package com.linagora.tmail.mailet;
+package com.linagora.tmail.common.chatlanguagemodel;
 
-import java.net.URL;
-import java.time.Duration;
-import java.util.Optional;
-
-import dev.langchain4j.model.chat.StreamingChatLanguageModel;
-import dev.langchain4j.model.openai.OpenAiStreamingChatModel;
-
-public class StreamChatLanguageModelFactory {
-    private static final String USE_DEFAULT_BASE_URL = "http://langchain4j.dev/demo/openai/v1";
-
-    public StreamingChatLanguageModel createChatLanguageModel(AIBotConfig config) {
-        String apiKey = config.getApiKey();
-        LlmModel llmModel = config.getLlmModel();
-        Optional<URL> baseURLOpt = config.getBaseURL();
-
-        return createOpenAILanguageModel(apiKey, llmModel.modelName(), baseURLOpt.map(URL::toString).orElse(USE_DEFAULT_BASE_URL),
-            config.getTimeout());
-    }
-
-    private StreamingChatLanguageModel createOpenAILanguageModel(String apiKey, String modelName, String baseUrl, Duration timeout) {
-        return OpenAiStreamingChatModel.builder()
-            .apiKey(apiKey)
-            .modelName(modelName)
-            .baseUrl(baseUrl)
-            .timeout(timeout)
-            .build();
-    }
+public record LlmModel(String modelName) {
 }
