+++ tmail-backend/tmail-third-party/ai-bot/src/test/java/com/linagora/tmail/jmap/aibot/AiBotCapabilitySaaSFilterIntegrationTest.java
@@ -37,14 +37,11 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
-import com.google.inject.Scopes;
 import com.google.inject.util.Modules;
 import com.linagora.tmail.james.app.MemoryConfiguration;
 import com.linagora.tmail.james.app.MemoryServer;
 import com.linagora.tmail.james.jmap.event.ApplyWhenFilter;
-import com.linagora.tmail.mailet.AIRedactionalHelper;
-import com.linagora.tmail.mailet.AIRedactionalHelperForTest;
-import com.linagora.tmail.mailet.conf.AIBaseModule;
+import com.linagora.tmail.common.AIBaseModule;
 import com.linagora.tmail.module.LinagoraTestJMAPServerModule;
 import com.linagora.tmail.saas.api.SaaSAccountRepository;
 import com.linagora.tmail.saas.api.memory.MemorySaaSAccountRepository;
