+++ tmail-backend/tmail-third-party/ai-bot/src/test/java/com/linagora/tmail/scribe/AIChatCompletionRoutesTest.java
@@ -46,7 +46,7 @@ import com.linagora.tmail.extension.WireMockAiServerExtension;
 import com.linagora.tmail.james.app.MemoryConfiguration;
 import com.linagora.tmail.james.app.MemoryServer;
 import com.linagora.tmail.james.jmap.event.ApplyWhenFilter;
-import com.linagora.tmail.mailet.conf.AIBaseModule;
+import com.linagora.tmail.common.AIBaseModule;
 import com.linagora.tmail.module.LinagoraTestJMAPServerModule;
 import com.linagora.tmail.saas.api.SaaSAccountRepository;
 import com.linagora.tmail.saas.api.memory.MemorySaaSAccountRepository;
