+++ tmail-backend/tmail-third-party/ai-bot/src/test/resources/extensions.properties
@@ -25,4 +25,4 @@
 # Upon injections, the user can reference additional guice modules, that are going to be used only upon extensions instantiation.
 
 #List of coma separated (',') fully qualified class names of additional guice modules to be used to instantiate extensions
-guice.extension.module=com.linagora.tmail.mailet.conf.AIBaseModule,com.linagora.tmail.mailet.conf.RagDeletionModule,com.linagora.tmail.jmap.aibot.AiMethodModule
+guice.extension.module=com.linagora.tmail.common.AIBaseModule,com.linagora.tmail.rag.RagDeletionModule,com.linagora.tmail.jmap.aibot.AiMethodModule
